requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Gemini
google-generativeai==0.3.0

# PDF Processing
pymupdf==1.23.8
pdfplumber==0.10.3
Pillow==10.1.0

# Vector Stores
chromadb==0.4.18
qdrant-client==1.6.4
pinecone-client==3.0.0

# Embeddings
sentence-transformers==2.2.2
openai==0.28.1
langchain

# Utilities
backoff==2.2.1
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
PyYAML==6.0.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
